Renaming getDefaultConfig -> createConfig
