+++ packages/kit/src/config/createConfig.ts
@@ -1,9 +1,10 @@
-import { createConfig, http, type CreateConfigParameters, type Config } from 'wagmi'
+import { createConfig as createWagmiConfig, type CreateConfigParameters, type Config } from 'wagmi'
 
 import { KitConfig, WalletType } from '../types'
 
 import { getDefaultChains } from './defaultChains'
 import { DefaultConnectorsProps, getDefaultConnectors } from './defaultConnectors'
+import { getDefaultTransports } from './defaultTransports'
 
 type DefaultConfigProps<T extends WalletType> = KitConfig &
   DefaultConnectorsProps<T> & {
@@ -16,11 +17,11 @@ export interface SequenceKitConfig {
   kitConfig: KitConfig
 }
 
-export const getDefaultConfig = <T extends WalletType>(walletType: T, props: DefaultConfigProps<T>): SequenceKitConfig => {
+export const createConfig = <T extends WalletType>(walletType: T, props: DefaultConfigProps<T>): SequenceKitConfig => {
   const { projectAccessKey, chainIds, wagmiConfig, ...rest } = props
 
   const chains = wagmiConfig?.chains || getDefaultChains(chainIds)
-  const transports = wagmiConfig?.transports || Object.fromEntries(chains.map(chain => [chain.id, http()]))
+  const transports = wagmiConfig?.transports || getDefaultTransports(chains)
   const connectors = wagmiConfig?.connectors || getDefaultConnectors(walletType, props)
 
   return {
@@ -28,7 +29,7 @@ export const getDefaultConfig = <T extends WalletType>(walletType: T, props: Def
       projectAccessKey,
       ...rest
     },
-    wagmiConfig: createConfig({
+    wagmiConfig: createWagmiConfig({
       ...wagmiConfig,
       chains,
       transports,
